Start moving layers / language functionality to OniEditor
