+++ browser/src/Editor/NeovimEditor/NeovimEditor.tsx
@@ -77,10 +77,6 @@ import { Rename } from "./Rename"
 import { Symbols } from "./Symbols"
 import { IToolTipsProvider, NeovimEditorToolTipsProvider } from "./ToolTipsProvider"
 
-// import { BufferScrollBarContainer } from "./containers/BufferScrollBarContainer"
-// import { DefinitionContainer } from "./containers/DefinitionContainer"
-// import { ErrorsContainer } from "./containers/ErrorsContainer"
-
 export class NeovimEditor extends Editor implements IEditor {
     private _bufferManager: BufferManager
     private _neovimInstance: NeovimInstance
@@ -133,6 +129,10 @@ export class NeovimEditor extends Editor implements IEditor {
         return this._neovimInstance
     }
 
+    public get bufferLayers(): BufferLayerManager {
+        return this._bufferLayerManager
+    }
+
     public get syntaxHighlighter(): ISyntaxHighlighter {
         return this._syntaxHighlighter
     }
@@ -155,27 +155,6 @@ export class NeovimEditor extends Editor implements IEditor {
         this._toolTipsProvider = new NeovimEditorToolTipsProvider(this._actions)
 
         this._bufferLayerManager = new BufferLayerManager()
-
-        // this._bufferLayerManager.addBufferLayer("*", (buf) => ({
-        //     id: "test",
-        //     render: (context) =>{
-        //         return context.isActive ?  <BufferScrollBarContainer /> : null
-        //     }
-        // }))
-
-        // this._bufferLayerManager.addBufferLayer("*", (buf) => ({
-        //     id: "test2",
-        //     render: (context) =>{
-        //         return context.isActive ?  <DefinitionContainer /> : null
-        //     }
-        // }))
-
-        // this._bufferLayerManager.addBufferLayer("*", (buf) => ({
-        //     id: "test3",
-        //     render: (context) =>{
-        //         return context.isActive ?  <ErrorsContainer /> : null
-        //     }
-        // }))
 
         this._contextMenuManager = new ContextMenuManager(this._toolTipsProvider, this._colors)
