+++ browser/src/Editor/OniEditor/containers/ErrorsContainer.ts
@@ -2,8 +2,8 @@ import { connect } from "react-redux"
 
 import { Errors, IErrorsProps } from "./../../../UI/components/Error"
 
-import * as Selectors from "./../NeovimEditorSelectors"
-import * as State from "./../NeovimEditorStore"
+import * as Selectors from "./../../NeovimEditor/NeovimEditorSelectors"
+import * as State from "./../../NeovimEditor/NeovimEditorStore"
 
 const mapStateToProps = (state: State.IState): IErrorsProps => {
